integer

+++ src/main/kotlin/org/polyfrost/crosshair/render/CrosshairRenderer.kt
@@ -97,7 +97,7 @@ object CrosshairRenderer {
         GL.scale(1 / mcScale, 1 / mcScale, 1f)
         val crosshair = ModConfig.newCurrentCrosshair
         GL.translate(crosshair.offsetX.toFloat(), crosshair.offsetY.toFloat(), 0f)
-        GL.translate(UResolution.windowWidth / 2f, UResolution.windowHeight / 2f, 0f)
+        GL.translate((UResolution.windowWidth / 2).toFloat(), (UResolution.windowHeight / 2).toFloat(), 0f)
         GL.rotate(crosshair.rotation.toFloat(), 0f, 0f, 1f)
         val scale = ModConfig.newCurrentCrosshair.scale / 100f
         val textureSize = if (ModConfig.mode) drawingImage.width else 15
